Get data from a SDMX API

Retrieve data from an sdmx api and return it as a data.frame

Usage

# S3 method for class 'sdmx_v2_1_data_request'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  format = c("csv", "json", "xml"),
  labels = c("both", "id"),
  ...,
  language = "nl",
  as.data.table = FALSE
)

Arguments

x

An httr2 request object

row.names

Not used

optional

Not used

format

The format of the data. Either "csv" or "json"

labels

The labels to include in the data. Either "both" or "id"

...

reserved for future use

language

The language to use for the text used in the response

as.data.table

if TRUE return a data.table::data.table().

Value

a data.frame or data.table depending on the value of as.data.table

Examples

if (interactive()){
  dfs <- get_dataflows()
  flowRef <- dfs$ref[2]

  system.time({
    d <-
      sdmx_v2_1_data_request(flowRef = flowRef)

    sdmx_v2_1_as_data_frame()
  })

  # json
  d <-
    sdmx_v2_1_data_request(flowRef = flowRef) |>
    sdmx_v2_1_as_data_frame(format = "json")

}