Get data from a SDMX API

Retrieve data from an sdmx api and return it as an xml_document

Usage

# S3 method for class 'sdmx_v2_1_structure_request'
as_xml_document(
  x,
  ...,
  cache_dir = tempdir(),
  verbose = getOption("cbsopendata.verbose", FALSE)
)

Arguments

x

An httr2 request object

...

reserved for future use

cache_dir

The directory to cache the request in, set to NULL to disable caching

verbose

if TRUE print information about the caching.

Value

a xml2::xml_document object

Examples

if (interactive()){
  dfs <- get_dataflows()
  flowRef <- dfs$ref[1]

  d <-
    sdmx_v2_1_structure_request(flowRef = flowRef) |>
    sdmx_v2_1_as_xml()

  d |>
    xml2::xml_find_all(".//generic:ObsValue") |>
    xml2::xml_text() |>
    head()

  path <- "example/dfs.json"
  d <-
    sdmx_v2_1_structure_request(
      detail = "referencestubs",
      # references = "all",
      format = "json"
    ) |>
    httr2::req_headers("Accept-Language" = "nl") |>
    httr2::req_perform(path=path)

}