Get sdmx dataflows

Get sdmx dataflows from a given endpoint.

Usage

get_dataflows(
  req = NULL,
  agencyID = NULL,
  ...,
  language = "nl",
  cache_dir = tempdir(),
  verbose = getOption("cbsopendata.verbose", FALSE)
)

Arguments

req

A character string for a given endpoint.

agencyID

A character string from a given agencyID.

...

saved for future use.

language

The language to use for the text used in the response.

cache_dir

The directory to cache the request in, set to NULL to disable caching.

verbose

if TRUE print information about the dataflows.

Value

a data.frame with available dataflows

Examples

dfs <- get_dataflows()
#> Available dataflows: 82
#> Agencies: ‘NL1’, ‘NL1.CNVT’, ‘NL1_DOUT’, ‘PT’
#> Content languages: ‘nl’, ‘en’

head(dfs)
#>               id version agencyID isFinal
#> 1 DF_TESTSET_X01     1.0      NL1   FALSE
#> 2 DF_TESTSET_X02     1.0      NL1   FALSE
#> 3    DF_24103NED     1.0 NL1.CNVT   FALSE
#> 4    DF_37230ned     1.0 NL1.CNVT   FALSE
#> 5       DF_37852     1.0 NL1.CNVT   FALSE
#> 6    DF_50012NED     1.0 NL1.CNVT   FALSE
#>                                                                       name
#> 1                                TESTSET (dataflow, only use for TESTSET!)
#> 2                                 TESTSET (dataflow, alleen voor TESTSET!)
#> 3                               Mbo; deelnemers, leerweg, opleiding, regio
#> 4                                  Bevolkingsontwikkeling; regio per maand
#> 5 Gezondheid, leefstijl, zorggebruik en -aanbod, doodsoorzaken; vanaf 1900
#> 6                              Geneesmiddelen; kosten en gebruik 2010-2015
#>                                                                   names.nl
#> 1                                TESTSET (dataflow, only use for TESTSET!)
#> 2                                 TESTSET (dataflow, alleen voor TESTSET!)
#> 3                               Mbo; deelnemers, leerweg, opleiding, regio
#> 4                                  Bevolkingsontwikkeling; regio per maand
#> 5 Gezondheid, leefstijl, zorggebruik en -aanbod, doodsoorzaken; vanaf 1900
#> 6                              Geneesmiddelen; kosten en gebruik 2010-2015
#>   names.en
#> 1     <NA>
#> 2     <NA>
#> 3     <NA>
#> 4     <NA>
#> 5     <NA>
#> 6     <NA>
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    description
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              t.b.v. testen .Stat applicaties
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                voor testen .Stat applicaties
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Deze tabel beschrijft het aantal personen dat op 1 oktober een inschrijving heeft aan de door de overheid bekostigde instellingen voor het middelbaar beroepsonderwijs (mbo) naar leerweg, opleiding (wel of niet zorg en welzijn), woonregio en studiejaar. Van de personen die een opleiding volgen binnen zorg en welzijn wordt aangegeven voor welke kwalificatie zij worden opgeleid. Deze tabel is ontwikkeld in het kader van het onderzoeksprogramma Arbeidsmarkt, Zorg en Welzijn (AZW), zie ook paragraaf 3.<br><br>Gegevens beschikbaar vanaf: 2010/'11.<br><br>Status van de cijfers:<br>De cijfers over 2010/'11 tot en met 2022/'23 zijn definitief en de cijfers over 2023/'24 zijn voorlopig.<br><br>Wijzigingen per 21 mei 2024:<br>De definitieve cijfers over 2022/'23 en de voorlopige cijfers over 2023/’24 zijn toegevoegd.<br><br>Alle lege cellen in de tabel zijn gevuld met een nul (0). Dit wijkt af van de richtlijnen beschreven in paragraaf 2. De reden is dat het niet altijd goed na te gaan is vanaf wanneer deelname aan een studie wel of niet mogelijk is. Het is daarmee ook niet goed mogelijk om te bepalen wanneer cijfers op logische gronden niet voor kunnen komen.<br><br>Vanaf 2022/'23 zijn dossiers persoonlijk begeleider zorgboerderij en werkbegeleider zorgbedrijf dierhouderij samengevoegd tot het nieuwe dossier zorgboerderij. De kwalificaties die vielen binnen deze dossiers zijn overgegaan naar het nieuwe dossier.<br><br>De kwalificaties zonder nadere differentiatie (z.n.d.) die vallen onder de volgende dossiers zijn uit de tabel verwijderd, omdat er op deze deze kwalificaties in de getoonde jaren geen inschrijvingen waren:<br>- Orthopedische techniek<br>- Orthopedische schoentechniek<br><br>Per 1 september 2022 is de RegioPlus-Arbeidsmarktregio Rijnstreek van naam gewijzigd: de nieuwe naam is “Rijn Gouwe”. Deze naamswijziging is in deze tabel teruggelegd in de tijd.<br><br>Wanneer komen er nieuwe cijfers?<br>In het tweede kwartaal van 2025 komen de definitieve cijfers over 2023/’24 en de voorlopige cijfers over 2024/'25 beschikbaar.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Deze tabel bevat cijfers over de veranderingen in de bevolkingsomvang van Nederland per regio. Deze veranderingen kunnen ontstaan door levend geboren kinderen, overledenen, immigratie, emigratie (inclusief administratieve correcties) en overige correcties. Daarnaast zijn ook cijfers over de totale bevolkingsgroei, begin- en eindbevolking opgenomen.<br><br>De in de tabel opgenomen regio's zijn landsdelen, provincies, COROP-gebieden en gemeenten. De gepresenteerde regiototalen betreffen samentellingen van gemeenten. In geval van grenswijzigingen die over verschillende regiogrenzen heen gaan is de indeling van de gemeenten gegroepeerd naar de meest recente situatie.<br><br>Gegevens beschikbaar vanaf: januari 2002.<br><br>Status van de cijfers:<br>De cijfers tot en met 2023 zijn definitief. De cijfers van de ‘Bevolking aan het begin van de periode’ voor januari 2024 zijn definitief.<br>De overige cijfers vanaf 2024 zijn voorlopig. Tussentijdse bijstellingen van voorgaande maanden zijn mogelijk.<br><br>Wijzigingen per 29 november 2024:<br>De voorlopige cijfers over oktober 2024 zijn toegevoegd.<br><br>Wanneer komen er nieuwe cijfers?<br>Aan het einde van iedere maand worden de voorlopige cijfers over de voorafgaande maand gepubliceerd. Tussentijdse bijstellingen van voorgaande maanden zijn mogelijk.<br>In het derde kwartaal van elk jaar worden de voorlopige cijfers over het voorgaande jaar vervangen door definitieve cijfers.
#> 5 Deze tabel toont de grote variëteit aan langlopende reeksen op het gebied  van gezondheid, leefstijl en gezondheidszorg. Cijfers over geboorte en  sterfte, enkele doodsoorzaken en het optreden van bepaalde infectieziekten zijn beschikbaar vanaf 1900. Andere reeksen starten op een later tijdstip.<br>Naast de ervaren gezondheid bevat de tabel onder meer cijfers over infectieziekten, ziekenhuisopnamen naar diagnose, levensverwachting, leefstijlfactoren als roken, alcohol en overgewicht, en doodsoorzaken. Ook diverse aspecten van de gezondheidszorg zoals het aantal beroepsbeoefenaren, het aantal beschikbare ziekenhuisbedden, de gemiddelde verpleegduur en de uitgaven aan zorg staan in de tabel.<br>Veel onderwerpen staan ook, in meer detail, in andere StatLine-tabellen, maar soms met een kortere looptijd. Gegevens over meldingsplichtige infectieziekten en Aids/hiv zijn niet in andere tabellen opgenomen.<br><br>Gegevens beschikbaar vanaf: 1900<br><br>Status van de cijfers:<br>2024:<br>De beschikbare cijfers zijn definitief.<br><br>2023:<br>De meeste beschikbare cijfers zijn definitief.<br>Cijfers zijn voorlopig voor:<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- uitgaven aan gezondheidszorg en welzijnszorg;<br>- perinatale sterfte (28 wk), zuigelingensterfte.<br>2022:<br>De meeste beschikbare cijfers zijn definitief.<br>Cijfers zijn voorlopig voor:<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- diagnoses bij ziekenhuisopname;<br>- ziekenhuisopnamen, verpleegdagen, verpleegduur;<br>- aantal ziekenhuisbedden;<br>- beroepen in de gezondheidszorg;<br>- uitgaven aan gezondheidszorg en welzijnszorg.<br>2021:<br>De meeste beschikbare cijfers zijn definitief.<br>Cijfers zijn voorlopig voor:<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- uitgaven aan gezondheidszorg en welzijnszorg.<br>2020 en eerder:<br>Meeste beschikbare cijfers zijn definitief.<br>Vanwege het dynamische karakter van de registratie zijn cijfers over alle jaren voorlopig voor meldingsplichtige infectieziekten, hiv, aids.<br><br><br>Wijzigingen per 18 december 2024:<br>- Vanwege een revisie van de statistieken Uitgaven gezondheids- en welzijnszorg 2021 zijn cijfers voor de uitgaven aan gezondheidszorg en welzijnszorg vanaf 2021 vervangen.<br>- Gereviseerde cijfers over de volume-index van de zorgkosten zijn nog niet beschikbaar, vanaf 2021 zijn deze cijfers leeggemaakt.<br>Aanvulling met de beschikbare meest recente cijfers:<br>- levendgeborenen, overledenen;<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- aantal ziekenhuisbedden;<br>- uitgaven aan gezondheidszorg en welzijnszorg;<br>- perinatale sterfte (28 wk), zuigelingensterfte;<br>- (gezonde) levensverwachting;<br>- doodsoorzaken.<br><br><br>Wanneer komen er nieuwe cijfers?<br>In juli 2025 verschijnen de op dat moment beschikbare meest recente cijfers.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   De cijfers in deze tabel zijn afkomstig uit het Genees- en hulpmiddelen Informatie Project (GIP geneesmiddelen) van het Zorginstituut Nederland (ZiN). Deze tabel is gemaakt voor De Staat van Volksgezondheid en Zorg en bevat gegevens over de ontwikkelingen in het gebruik van geneesmiddelen.<br>Het betreft zorg die extramuraal (d.w.z. buiten instellingen als ziekenhuizen en verpleeghuizen) wordt verleend op basis van de Zorgverzekeringswet. Dus alleen de extramuraal verstrekte geneesmiddelen vergoed uit de verplichte basisverzekering voor geneeskundige zorg.<br>Het gebruik van geneesmiddelen kan worden uitgesplitst naar: verzekerdenkenmerken (leeftijd en geslacht), soort geneesmiddel (ATC-classificatie). Outputvariabelen zijn totale kosten en aantallen gebruikers.<br>Deze tabel is stopgezet, zie in paragraaf 4 de link naar de GIP databank. Daar zijn de cijfers ook beschikbaar.<br><br>Gegevens beschikbaar van 2010 tot en met 2015.<br><br>Status van de cijfers:<br>Jaren 2010-2013 zijn definitief, 2014 en 2015 zijn voorlopig. Aangezien deze tabel is stopgezet, worden de gegevens niet meer definitief gemaakt.<br><br>Wijzigingen per 08-08-2019:<br>Geen, deze tabel is stopgezet.<br><br>Wanneer komen er nieuwe cijfers?<br>Niet meer van toepassing.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             nl
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              t.b.v. testen .Stat applicaties
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                voor testen .Stat applicaties
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     Deze tabel beschrijft het aantal personen dat op 1 oktober een inschrijving heeft aan de door de overheid bekostigde instellingen voor het middelbaar beroepsonderwijs (mbo) naar leerweg, opleiding (wel of niet zorg en welzijn), woonregio en studiejaar. Van de personen die een opleiding volgen binnen zorg en welzijn wordt aangegeven voor welke kwalificatie zij worden opgeleid. Deze tabel is ontwikkeld in het kader van het onderzoeksprogramma Arbeidsmarkt, Zorg en Welzijn (AZW), zie ook paragraaf 3.<br><br>Gegevens beschikbaar vanaf: 2010/'11.<br><br>Status van de cijfers:<br>De cijfers over 2010/'11 tot en met 2022/'23 zijn definitief en de cijfers over 2023/'24 zijn voorlopig.<br><br>Wijzigingen per 21 mei 2024:<br>De definitieve cijfers over 2022/'23 en de voorlopige cijfers over 2023/’24 zijn toegevoegd.<br><br>Alle lege cellen in de tabel zijn gevuld met een nul (0). Dit wijkt af van de richtlijnen beschreven in paragraaf 2. De reden is dat het niet altijd goed na te gaan is vanaf wanneer deelname aan een studie wel of niet mogelijk is. Het is daarmee ook niet goed mogelijk om te bepalen wanneer cijfers op logische gronden niet voor kunnen komen.<br><br>Vanaf 2022/'23 zijn dossiers persoonlijk begeleider zorgboerderij en werkbegeleider zorgbedrijf dierhouderij samengevoegd tot het nieuwe dossier zorgboerderij. De kwalificaties die vielen binnen deze dossiers zijn overgegaan naar het nieuwe dossier.<br><br>De kwalificaties zonder nadere differentiatie (z.n.d.) die vallen onder de volgende dossiers zijn uit de tabel verwijderd, omdat er op deze deze kwalificaties in de getoonde jaren geen inschrijvingen waren:<br>- Orthopedische techniek<br>- Orthopedische schoentechniek<br><br>Per 1 september 2022 is de RegioPlus-Arbeidsmarktregio Rijnstreek van naam gewijzigd: de nieuwe naam is “Rijn Gouwe”. Deze naamswijziging is in deze tabel teruggelegd in de tijd.<br><br>Wanneer komen er nieuwe cijfers?<br>In het tweede kwartaal van 2025 komen de definitieve cijfers over 2023/’24 en de voorlopige cijfers over 2024/'25 beschikbaar.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Deze tabel bevat cijfers over de veranderingen in de bevolkingsomvang van Nederland per regio. Deze veranderingen kunnen ontstaan door levend geboren kinderen, overledenen, immigratie, emigratie (inclusief administratieve correcties) en overige correcties. Daarnaast zijn ook cijfers over de totale bevolkingsgroei, begin- en eindbevolking opgenomen.<br><br>De in de tabel opgenomen regio's zijn landsdelen, provincies, COROP-gebieden en gemeenten. De gepresenteerde regiototalen betreffen samentellingen van gemeenten. In geval van grenswijzigingen die over verschillende regiogrenzen heen gaan is de indeling van de gemeenten gegroepeerd naar de meest recente situatie.<br><br>Gegevens beschikbaar vanaf: januari 2002.<br><br>Status van de cijfers:<br>De cijfers tot en met 2023 zijn definitief. De cijfers van de ‘Bevolking aan het begin van de periode’ voor januari 2024 zijn definitief.<br>De overige cijfers vanaf 2024 zijn voorlopig. Tussentijdse bijstellingen van voorgaande maanden zijn mogelijk.<br><br>Wijzigingen per 29 november 2024:<br>De voorlopige cijfers over oktober 2024 zijn toegevoegd.<br><br>Wanneer komen er nieuwe cijfers?<br>Aan het einde van iedere maand worden de voorlopige cijfers over de voorafgaande maand gepubliceerd. Tussentijdse bijstellingen van voorgaande maanden zijn mogelijk.<br>In het derde kwartaal van elk jaar worden de voorlopige cijfers over het voorgaande jaar vervangen door definitieve cijfers.
#> 5 Deze tabel toont de grote variëteit aan langlopende reeksen op het gebied  van gezondheid, leefstijl en gezondheidszorg. Cijfers over geboorte en  sterfte, enkele doodsoorzaken en het optreden van bepaalde infectieziekten zijn beschikbaar vanaf 1900. Andere reeksen starten op een later tijdstip.<br>Naast de ervaren gezondheid bevat de tabel onder meer cijfers over infectieziekten, ziekenhuisopnamen naar diagnose, levensverwachting, leefstijlfactoren als roken, alcohol en overgewicht, en doodsoorzaken. Ook diverse aspecten van de gezondheidszorg zoals het aantal beroepsbeoefenaren, het aantal beschikbare ziekenhuisbedden, de gemiddelde verpleegduur en de uitgaven aan zorg staan in de tabel.<br>Veel onderwerpen staan ook, in meer detail, in andere StatLine-tabellen, maar soms met een kortere looptijd. Gegevens over meldingsplichtige infectieziekten en Aids/hiv zijn niet in andere tabellen opgenomen.<br><br>Gegevens beschikbaar vanaf: 1900<br><br>Status van de cijfers:<br>2024:<br>De beschikbare cijfers zijn definitief.<br><br>2023:<br>De meeste beschikbare cijfers zijn definitief.<br>Cijfers zijn voorlopig voor:<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- uitgaven aan gezondheidszorg en welzijnszorg;<br>- perinatale sterfte (28 wk), zuigelingensterfte.<br>2022:<br>De meeste beschikbare cijfers zijn definitief.<br>Cijfers zijn voorlopig voor:<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- diagnoses bij ziekenhuisopname;<br>- ziekenhuisopnamen, verpleegdagen, verpleegduur;<br>- aantal ziekenhuisbedden;<br>- beroepen in de gezondheidszorg;<br>- uitgaven aan gezondheidszorg en welzijnszorg.<br>2021:<br>De meeste beschikbare cijfers zijn definitief.<br>Cijfers zijn voorlopig voor:<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- uitgaven aan gezondheidszorg en welzijnszorg.<br>2020 en eerder:<br>Meeste beschikbare cijfers zijn definitief.<br>Vanwege het dynamische karakter van de registratie zijn cijfers over alle jaren voorlopig voor meldingsplichtige infectieziekten, hiv, aids.<br><br><br>Wijzigingen per 18 december 2024:<br>- Vanwege een revisie van de statistieken Uitgaven gezondheids- en welzijnszorg 2021 zijn cijfers voor de uitgaven aan gezondheidszorg en welzijnszorg vanaf 2021 vervangen.<br>- Gereviseerde cijfers over de volume-index van de zorgkosten zijn nog niet beschikbaar, vanaf 2021 zijn deze cijfers leeggemaakt.<br>Aanvulling met de beschikbare meest recente cijfers:<br>- levendgeborenen, overledenen;<br>- meldingsplichtige infectieziekten, hiv, aids;<br>- aantal ziekenhuisbedden;<br>- uitgaven aan gezondheidszorg en welzijnszorg;<br>- perinatale sterfte (28 wk), zuigelingensterfte;<br>- (gezonde) levensverwachting;<br>- doodsoorzaken.<br><br><br>Wanneer komen er nieuwe cijfers?<br>In juli 2025 verschijnen de op dat moment beschikbare meest recente cijfers.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   De cijfers in deze tabel zijn afkomstig uit het Genees- en hulpmiddelen Informatie Project (GIP geneesmiddelen) van het Zorginstituut Nederland (ZiN). Deze tabel is gemaakt voor De Staat van Volksgezondheid en Zorg en bevat gegevens over de ontwikkelingen in het gebruik van geneesmiddelen.<br>Het betreft zorg die extramuraal (d.w.z. buiten instellingen als ziekenhuizen en verpleeghuizen) wordt verleend op basis van de Zorgverzekeringswet. Dus alleen de extramuraal verstrekte geneesmiddelen vergoed uit de verplichte basisverzekering voor geneeskundige zorg.<br>Het gebruik van geneesmiddelen kan worden uitgesplitst naar: verzekerdenkenmerken (leeftijd en geslacht), soort geneesmiddel (ATC-classificatie). Outputvariabelen zijn totale kosten en aantallen gebruikers.<br>Deze tabel is stopgezet, zie in paragraaf 4 de link naar de GIP databank. Daar zijn de cijfers ook beschikbaar.<br><br>Gegevens beschikbaar van 2010 tot en met 2015.<br><br>Status van de cijfers:<br>Jaren 2010-2013 zijn definitief, 2014 en 2015 zijn voorlopig. Aangezien deze tabel is stopgezet, worden de gegevens niet meer definitief gemaakt.<br><br>Wijzigingen per 08-08-2019:<br>Geen, deze tabel is stopgezet.<br><br>Wanneer komen er nieuwe cijfers?<br>Niet meer van toepassing.
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                annotations
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    NonProductionDataflow
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    NonProductionDataflow
#> 3                                                                                             NonProductionDataflow, LAYOUT_COLUMN, LAYOUT_ROW, LAYOUT_ROWSECTION, DEFAULT, COMBINED_CONCEPTS, ALWAYS_DISPLAY_PARENTS, LAYOUT_FLAG, MAX_TABLE_DATA, NOT_DISPLAYED, LAYOUT_CELL_ALIGN, NA, Leerweg, Topics,Opleiding, Perioden,RegioS, Perioden=2023SJ00,RegioS=NL00,Topics=DeelnemersMbo_1, COMBINED_UNIT_MEASURE:UNIT_MEASURE,UNIT_MULT, Topics, OBS_STATUS, 10000, OBS_STATUS=D,UNIT_MULT=0, RIGHT, NA, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA
#> 4                                                                                 NonProductionDataflow, LAYOUT_COLUMN, LAYOUT_ROW, DEFAULT, COMBINED_CONCEPTS, ALWAYS_DISPLAY_PARENTS, LAYOUT_FLAG, MAX_TABLE_DATA, NOT_DISPLAYED, LAYOUT_CELL_ALIGN, NA, Perioden, RegioS,Topics, Perioden=2024MM01+2024MM02+2024MM03+2024MM04+2024MM05+2024MM06+2024MM07+2024MM08+2024MM09+2024MM10,RegioS=NL01, COMBINED_UNIT_MEASURE:UNIT_MEASURE,UNIT_MULT, Topics, OBS_STATUS, 10000, OBS_STATUS=D,UNIT_MULT=0, RIGHT, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA
#> 5                                                                                           NonProductionDataflow, LAYOUT_COLUMN, LAYOUT_ROW, DEFAULT, COMBINED_CONCEPTS, ALWAYS_DISPLAY_PARENTS, LAYOUT_FLAG, MAX_TABLE_DATA, NOT_DISPLAYED, LAYOUT_CELL_ALIGN, NA, Perioden, Topics, Perioden=1900JJ00+1940JJ00+1980JJ00+1990JJ00+2000JJ00+2010JJ00+2020JJ00+2021JJ00+2022JJ00+2023JJ00+2024JJ00, COMBINED_UNIT_MEASURE:UNIT_MEASURE,UNIT_MULT, Topics, OBS_STATUS, 10000, OBS_STATUS=D,UNIT_MULT=0, RIGHT, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA
#> 6 NonProductionDataflow, LAYOUT_COLUMN, LAYOUT_ROW, LAYOUT_ROWSECTION, DEFAULT, COMBINED_CONCEPTS, ALWAYS_DISPLAY_PARENTS, LAYOUT_FLAG, MAX_TABLE_DATA, NOT_DISPLAYED, LAYOUT_CELL_ALIGN, NA, Perioden, Topics,Geneesmiddelengroep, Geslacht,Leeftijd, Geslacht=T001038,Leeftijd=10000,Perioden=2010JJ00+2011JJ00+2012JJ00+2013JJ00+2014JJ00+2015JJ00,Topics=GebruikersVanGeneesmiddelen_1, COMBINED_UNIT_MEASURE:UNIT_MEASURE,UNIT_MULT, Topics, OBS_STATUS, 10000, OBS_STATUS=D,UNIT_MULT=0, RIGHT, NA, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, COMBINED_UNIT_MEASURE:Meeteenheid, NA, NA, NA, NA, NA
#>                                                                            structure
#> 1   urn:sdmx:org.sdmx.infomodel.datastructure.DataStructure=NL1:DSD_TESTSET_X01(1.0)
#> 2   urn:sdmx:org.sdmx.infomodel.datastructure.DataStructure=NL1:DSD_TESTSET_X02(1.0)
#> 3 urn:sdmx:org.sdmx.infomodel.datastructure.DataStructure=NL1.CNVT:DSD_24103NED(1.0)
#> 4 urn:sdmx:org.sdmx.infomodel.datastructure.DataStructure=NL1.CNVT:DSD_37230ned(1.0)
#> 5    urn:sdmx:org.sdmx.infomodel.datastructure.DataStructure=NL1.CNVT:DSD_37852(1.0)
#> 6 urn:sdmx:org.sdmx.infomodel.datastructure.DataStructure=NL1.CNVT:DSD_50012NED(1.0)
#>                        ref                  flowRef
#> 1   NL1,DF_TESTSET_X01,1.0   NL1,DF_TESTSET_X01,1.0
#> 2   NL1,DF_TESTSET_X02,1.0   NL1,DF_TESTSET_X02,1.0
#> 3 NL1.CNVT,DF_24103NED,1.0 NL1.CNVT,DF_24103NED,1.0
#> 4 NL1.CNVT,DF_37230ned,1.0 NL1.CNVT,DF_37230ned,1.0
#> 5    NL1.CNVT,DF_37852,1.0    NL1.CNVT,DF_37852,1.0
#> 6 NL1.CNVT,DF_50012NED,1.0 NL1.CNVT,DF_50012NED,1.0